pyproject.toml
[project]
name = "titanic-mlops"
version = "0.1.0"
description = "Lab 0 - Titanic Training Pipeline | ITI MLOps Course"
authors = [{ name = "Aly Ayman", email = "alyayman2020@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.3.0",
    "loguru>=0.7.2",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.4.0",
]

[dependency-groups]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ── Black ──────────────────────────────────────────────────────────────────────
[tool.black]
line-length = 99
target-version = ["py311"]

# ── isort ─────────────────────────────────────────────────────────────────────
[tool.isort]
profile = "black"
line_length = 99
known_first_party = ["src"]

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 99
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
extend-select = ["I"]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true

# ── pytest ─────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
